base file for visualization
